// src/services/api/generated/model/ingredientNutritionProperty.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * A Cheffy API é o backend central de uma aplicação de receitas, projetado para gerenciar catálogo culinário, categorias, tags, ingredientes, informações nutricionais, favoritos de usuário, autenticação via Better Auth, assinatura de upload no Cloudinary e um assistente gastronômico integrado à Groq.
 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type IngredientNutritionProperty = {
  id: string;
  ingredientId: string;
  energyKcalPer100g: number;
  carbohydratesPer100g: number;
  /** @nullable */
  totalSugarsPer100g: number | null;
  /** @nullable */
  addedSugarsPer100g: number | null;
  proteinPer100g: number;
  totalFatPer100g: number;
  /** @nullable */
  saturatedFatPer100g: number | null;
  /** @nullable */
  transFatPer100g: number | null;
  /** @nullable */
  fiberPer100g: number | null;
  /** @nullable */
  sodiumMgPer100g: number | null;
  /** @nullable */
  source: string | null;
} | null;